servers/docker-python/requirements.txt
python-socketio==4.4.0
aiohttp==3.6.2
kafka
kafka-python
pymongo
